requirements.txt
pydantic
pymupdf
openai
qdrant-client
azure-search-documents
azure-core
python-dotenv
